+++ src/os2datascanner/engine2/model/smbc.py
@@ -398,6 +398,10 @@ class SMBCHandle(Handle):
     def presentation_url(self):
         return make_presentation_url(self)
 
+    @property
+    def container_url(self):
+        return self.presentation_url.rsplit('/', maxsplit=1)[0]
+
     def __str__(self):
         return make_full_windows_path(self)